Mise a jour du UI
modification du MainActivity.java, activity_main.xml, styles.xml
ajout du SendActionListener.java, layouthomepage.xml
