+++ app/src/main/java/sam/ccnb/chiffrementmobile/MainActivity.java
@@ -1,8 +1,11 @@
 /*  Programmeur: Samuel Lavoie, Jean-Philippe Haché
     Date:   Avril, 2015
 
-    Dernière mise a jour;
+    Première mise a jour;
         - (Samuel Lavoie, 28 Avril, 2015) implémentation du JAR (pour l'API du chifremment)
+
+    Dernière mise a jour;
+        -(Julien Guy Losier, 5 mai, 2015) changer l'application de départ du projet
  */
 
 package sam.ccnb.chiffrementmobile;
@@ -11,45 +14,19 @@ import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
 
-import sam.ccnb.chiffrementmobile.KeyRsaHandler;
-import sam.ccnb.chiffrementmobile.ChiffrementHandler;
-
-
-public class MainActivity extends ActionBarActivity implements View.OnClickListener{
-
-    private Button sendButton;
-    private TextView receiver;
-    private TextView output;
-    private EditText message;
+public class MainActivity extends ActionBarActivity{
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        receiver = (TextView) findViewById(R.id.receiverNameTextView);
-        output = (TextView) findViewById(R.id.outputTextView);
-
-        message = (EditText) findViewById(R.id.sendEditText);
-
-        sendButton = (Button) findViewById(R.id.sendButton);
-        sendButton.setOnClickListener(this);
-
+        setContentView(R.layout.layouthomepage);
     }
-
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.menu_main, menu);
-
-
-
         return true;
     }
 
@@ -68,34 +45,6 @@ public class MainActivity extends ActionBarActivity implements View.OnClickListe
         return super.onOptionsItemSelected(item);
     }
 
-    @Override
-    public void onClick(View v) {
 
-        String nom, msg;
-
-        //Placer les values des Textfields dans un string
-        nom = receiver.getText().toString();
-        msg = message.getText().toString();
-
-        //Dechifremment du message
-        //String dechMsg = ChiffrementHandler.decrypt( chiMsg, kh.getPriKey() );
-        //reMsgTxt.setText( dechMsg );
-
-        byte[] chiMessage = null;
-        String received = null;
-
-        //loader le KeyHandler
-        KeyRsaHandler key = new KeyRsaHandler(nom.hashCode());
-
-        //chiffrer le message, retourne un array de char
-        chiMessage = ChiffrementHandler.encrypt(msg, key.getPubKey());
-
-        //afficher le message brut
-        String bufferString = new String( chiMessage );
-        output.setText("Message brut: " + bufferString);
-
-        //Dechifremment du message
-        received = ChiffrementHandler.decrypt(chiMessage, key.getPriKey());
-        output.append(" Message Decrypter: " + received);
-    }
 }
+
