+++ app/src/main/res/layout/activity_main.xml
@@ -92,16 +92,13 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Receiver name"
+        android:text="New Text"
         android:id="@+id/receiverNameTextView"
-        android:layout_alignRight="@+id/outputTextView"
-        android:layout_alignEnd="@+id/outputTextView"
-        android:layout_below="@+id/logoImageView"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:layout_above="@+id/outputTextView"
-        android:gravity="center"
-        android:editable="false"
-        android:textColor="#ffffffff" />
+        android:layout_below="@+id/logoImageView"
+        android:layout_alignRight="@+id/outputTextView"
+        android:layout_alignEnd="@+id/outputTextView"
+        android:layout_above="@+id/outputTextView" />
 
 </RelativeLayout>
